+++ jtorch/src/jtorch/spatial_lp_pooling.cpp
@@ -93,7 +93,7 @@ namespace jtorch {
 
     // Wait for all threads to finish
     std::unique_lock<std::mutex> ul(thread_update_lock_);  // Get lock
-    while (threads_finished_ != thread_cbs_->size()) {
+    while (threads_finished_ != static_cast<int32_t>(thread_cbs_->size())) {
       not_finished_.wait(ul);
     }
     ul.unlock();  // Release lock
